login bug fixed

+++ frontend/src/api.js
@@ -12,12 +12,25 @@ api.interceptors.request.use(config => {
 });
 
 // 401 自动跳登录
+// api.interceptors.response.use(
+//   res => res,
+//   err => {
+//     if (err.response?.status === 401) {
+//       localStorage.clear();
+//       window.location.href = '/login';
+//     }
+//     return Promise.reject(err);
+//   }
+// );
+
 api.interceptors.response.use(
   res => res,
   err => {
     if (err.response?.status === 401) {
-      localStorage.clear();
-      window.location.href = '/login';
+      if (!window.location.pathname.includes('/login')) {
+        localStorage.clear();
+        window.location.href = '/login';
+      }
     }
     return Promise.reject(err);
   }
